Move Credentials and UsernamePassword to auth package
