+++ core/src/test/java/org/apache/brooklyn/util/executor/HttpExecutorImplTest.java
@@ -29,12 +29,11 @@ import com.google.common.base.Function;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Collections2;
 import org.apache.brooklyn.util.core.http.BetterMockWebServer;
-import org.apache.brooklyn.util.executor.HttpExecutorFactoryImpl;
 import org.apache.brooklyn.util.http.executor.HttpExecutor;
 import org.apache.brooklyn.util.http.executor.HttpRequest;
 import org.apache.brooklyn.util.http.executor.HttpRequestImpl;
 import org.apache.brooklyn.util.http.executor.HttpResponse;
-import org.apache.brooklyn.util.http.executor.UsernamePassword;
+import org.apache.brooklyn.util.http.auth.UsernamePassword;
 import org.apache.commons.codec.binary.Base64;
 import org.bouncycastle.util.Strings;
 import org.testng.annotations.AfterMethod;
